fix: address review comments — stack alloc, remove Flush, normalize status codes, shared helper
- Use stack-allocated [8]byte array in generateRequestID to avoid heap alloc
- Remove Flush() method from responseWriter; Unwrap() enables
  http.ResponseController to detect real flush support
- Normalize gRPC status_code to int (matching HTTP) for consistent log querying
- Move testServerStream to helpers_test.go shared across test files
